+++ lib/epages/rest/api.rb
@@ -8,6 +8,7 @@ require 'epages/rest/orders'
 require 'epages/rest/products'
 require 'epages/rest/sales'
 require 'epages/rest/shipping_methods'
+require 'epages/rest/tax_classes'
 
 module Epages
   module REST
@@ -22,6 +23,7 @@ module Epages
       include REST::Products
       include REST::Sales
       include REST::ShippingMethods
+      include REST::TaxClasses
     end
   end
 end