+++ lib/epages/rest/carts.rb
@@ -72,6 +72,19 @@ module Epages
         id = epages_id(cart)
         perform_post_with_object("/carts/#{id}/order", {}, Epages::Order)
       end
+
+      # implements the call https://developer.epages.com/apps/api-reference/post-shops-shopid-carts-cartid-coupon.html
+      def apply_coupon(cart, data = {})
+        id = epages_id(cart)
+        perform_post_with_object("/carts/#{id}/coupon", data, Epages::Cart)
+      end
+
+      # implements the call https://developer.epages.com/apps/api-reference/delete-shops-shopid-carts-cartid-coupon-couponlineitemid.html
+      # It needs the couponLineItemId to be able to delete it
+      def delete_coupon(cart, coupon_line_item_id)
+        id = epages_id(cart)
+        perform_delete_with_object("/carts/#{id}/coupon/#{coupon_line_item_id}", {}, Epages::Cart)
+      end
     end
   end
 end